requirements.txt
Flask
Pillow
pytesseract
opencv-python
numpy
rapidfuzz